Place Sapling circuit into submodule.
